refactor(checkboxes): moves checkboxes to new folder
