+++ sugar/tailwind-plugin-form/src/checkboxes_and_radios/checkbox_and_radio.plugin.ts
@@ -1,0 +1,12 @@
+import { SugarPluginComponent } from "@singlestone/tailwind-helpers-sugar";
+
+import { checkboxStyles } from "./checkbox.styles";
+
+export const sugarCheckboxAndRadio: SugarPluginComponent = (
+  { addComponents, theme },
+  { prefix }
+) => {
+  addComponents({
+    [`.${prefix}-checkbox`]: checkboxStyles(theme),
+  });
+};
