+++ sugar/tailwind-plugin-form/src/inputs/inputs.plugin.ts
@@ -1,6 +1,5 @@
 import { SugarPluginComponent } from "@singlestone/tailwind-helpers-sugar";
 
-import { checkboxStyles } from "./checkbox_and_radio.styles";
 import { textInput } from "./inputs.styles";
 import { selectInputStyles } from "./select-input.styles";
 
@@ -11,6 +10,5 @@ export const sugarInputs: SugarPluginComponent = (
   addComponents({
     [`.${prefix}-text-input`]: textInput(theme),
     [`.${prefix}-select-input`]: selectInputStyles(theme),
-    [`.${prefix}-checkbox`]: checkboxStyles(theme),
   });
 };
